Исправлен тест для подсчета очков
