+++ js/data/game.js
@@ -1,16 +1,6 @@
 import {changeScreen} from "../util";
+import {GAME_SETTINGS} from "./game-data";
 import gameScreen from "../game/game-screen";
-
-export const GAME_SETTINGS = {
-  totalQuestions: 10,
-  totalTime: 300000, // 5 min in ms
-  fastAnswerTime: 30000, // 30 sec in ms
-  maxMistakes: 3,
-  losePoints: -1,
-  regularPoints: 1,
-  fastPoints: 2,
-  wrongPoints: -2
-};
 
 export const INITIAL_STATE = Object.freeze({
   level: 0,
