+++ js/game/calculate-points.test.js
@@ -9,7 +9,7 @@ describe(`Check points calculator`, () => {
       {correct: true, time: 60000},
       {correct: true, time: 60000},
       {correct: true, time: 60000}
-    ]}), -1);
+    ], mistakes: 0}), -1);
   });
 
   it(`should return -1 if used all attempts`, () => {
@@ -20,11 +20,11 @@ describe(`Check points calculator`, () => {
       {correct: true, time: 60000},
       {correct: false, time: 60000},
       {correct: true, time: 60000},
+      {correct: true, time: 60000},
+      {correct: true, time: 60000},
       {correct: false, time: 60000},
-      {correct: true, time: 60000},
-      {correct: true, time: 60000},
       {correct: true, time: 60000}
-    ]}), -1);
+    ], mistakes: 3}), -1);
     assert.equal(calculatePoints({answers: [
       {correct: true, time: 10000},
       {correct: true, time: 10000},
@@ -33,10 +33,10 @@ describe(`Check points calculator`, () => {
       {correct: true, time: 10000},
       {correct: true, time: 10000},
       {correct: true, time: 10000},
-      {correct: true, time: 10000},
+      {correct: false, time: 10000},
       {correct: false, time: 10000},
       {correct: false, time: 10000}
-    ]}), -1);
+    ], mistakes: 3}), -1);
   });
 
   it(`should return 10 if all answers are correct and slow`, () => {
@@ -51,7 +51,7 @@ describe(`Check points calculator`, () => {
       {correct: true, time: 60000},
       {correct: true, time: 60000},
       {correct: true, time: 60000}
-    ]}), 10);
+    ], mistakes: 0}), 10);
   });
 
   it(`should return 20 if all answers are correct and fast`, () => {
@@ -66,7 +66,7 @@ describe(`Check points calculator`, () => {
       {correct: true, time: 20000},
       {correct: true, time: 20000},
       {correct: true, time: 20000}
-    ]}), 20);
+    ], mistakes: 0}), 20);
   });
 
   it(`should return between 0 and 20`, () => {
@@ -81,6 +81,6 @@ describe(`Check points calculator`, () => {
       {correct: true, time: 30000},
       {correct: true, time: 40000},
       {correct: true, time: 50000}
-    ]}), /^([0-1]?[0-9]|20)$/);
+    ], mistakes: 2}), /^([0-1]?[0-9]|20)$/);
   });
 });
